Cache activation rows in source dtype
Stop upcasting captured activation rows to fp32 at capture time; every
consumer already upcasts at its own math site. Halves the resident size
of the calibration row caches for bf16/fp16 models.

+++ src/diffuse_compressor/calibration/cache.py
@@ -51,7 +51,9 @@ class TensorCache:
             remaining = max_rows - self.num_rows
             if rows.shape[0] > remaining:
                 rows = rows[:remaining]
-        self.data.append(rows.float().cpu())
+        # Keep the source dtype: bf16/fp16 activations double in size when stored as
+        # fp32, and every consumer upcasts to fp32 at its own math site anyway.
+        self.data.append(rows.cpu())
         self.num_rows += rows.shape[0]
 
     def tensor(self) -> torch.Tensor | None:
